requirements.txt
# Core dependencies
flask==2.3.3
werkzeug==2.3.7
pillow==10.0.1
pdf2image==1.16.3
ollama==0.1.5

# LangGraph and LangChain dependencies
langgraph==0.0.24
langchain-core==0.1.16
langchain-ollama==0.0.7

# Additional dependencies
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2